backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
asyncpg==0.30.0
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.10

# Vector Database & Embeddings
pgvector==0.4.1
sentence-transformers==3.3.1
numpy==1.26.4
scikit-learn==1.5.2
datasets==3.2.0

# LLM Integrations
openai==1.57.4
anthropic==0.42.0
httpx==0.27.0
ollama==0.4.5

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
email-validator==2.2.0

# Redis & Caching
redis==5.2.1
aiocache==0.12.3

# NLP & Text Processing
spacy==3.8.3
jinja2==3.1.5
tiktoken==0.8.0

# Utilities
python-dotenv==1.0.1
click==8.1.8
rich==13.9.4
loguru==0.7.3

# Monitoring & Error Tracking
sentry-sdk[fastapi]==2.19.2
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1

# Development
black==25.1.0
ruff==0.9.2
mypy==1.14.0
pre-commit==4.0.1
